Initial commit: WGPU function grapher
This contains the main function and event loop to handle rendering
and user input; structures to hold the winit window and wgpu device
state, as well as a camera state; code for creating a render pipeline,
including a shader for rendering a solid-color mesh; some code for
creating and managing matrix uniforms.

Next we will add function for representing the (buffer, index, etc.)
data of a mesh, and we will add code render such meshes and code to
create and manage the mesh data that is to be rendered.