Create, copy, modify code towards creating state to render textures
Includes code to create and load a texture from an image, and starts
adding code to create scenes with textured meshes. The latter parts
are copies from the solid color mesh code, and still need many
modifications for textured use. Adds a vector of textured mesh data
structs to the Scene struct.