Add heat eqn; improve timing; better wave eqn stim.; misc improvements
The more continuous wave equation stimulus reduces the spurious
oscillations in the numerical solution.

There was an issue with the render of the heat equation near the
boundary, so we solve it on the whole domain but only render the
part some distance inside of the boundary. We'll look into this
further.

We use an accumulator to better control the framerate following
the example in (excellent) Pomona College lecture notes here:

 https://cs.pomona.edu/classes/cs181g/notes/controlling-time.html