Basic integration of wgpu_grapher with egui
This just has the basic support for rendering the 3D graph on the
same window as the GUI. Many other features need to be implemented,
and things could definitely be cleaned up and restructured. But the
basic functionality works.

Note: We need to figure out the framerate on Linux at least, because
it is causing high CPU utilization and producing too much heat.